desktop-scanner/requirements.txt
opencv-python
mysql-connector-python
pyzbar
numpy
playsound
